Replaced yahoo currency with ecb currency.

+++ upload/admin/model/localisation/currency.php
@@ -105,54 +105,58 @@ class ModelLocalisationCurrency extends Model {
 		}
 	}
 
-	public function refresh($force = false) {
-		$currency_data = array();
-
-		if ($force) {
-			$query = $this->db->query("SELECT * FROM `" . DB_PREFIX . "currency` WHERE `code` != '" . $this->db->escape($this->config->get('config_currency')) . "'");
-		} else {
-			$query = $this->db->query("SELECT * FROM `" . DB_PREFIX . "currency` WHERE `code` != '" . $this->db->escape($this->config->get('config_currency')) . "' AND `date_modified` < '" . $this->db->escape(date('Y-m-d H:i:s', strtotime('-1 day'))) . "'");
-		}
-
-		foreach ($query->rows as $result) {
-			$currency_data[] = $this->config->get('config_currency') . $result['code'] . '=X';
-			$currency_data[] = $result['code'] . $this->config->get('config_currency') . '=X';
-		}
-
+	public function refresh() {
 		$curl = curl_init();
 
-		curl_setopt($curl, CURLOPT_URL, 'http://download.finance.yahoo.com/d/quotes.csv?s=' . implode(',', $currency_data) . '&f=sl1&e=.json');
+		curl_setopt($curl, CURLOPT_URL, 'https://www.ecb.europa.eu/stats/eurofxref/eurofxref-daily.xml');
 		curl_setopt($curl, CURLOPT_RETURNTRANSFER, 1);
 		curl_setopt($curl, CURLOPT_HEADER, false);
+		curl_setopt($curl, CURLOPT_SSL_VERIFYPEER, 0);
 		curl_setopt($curl, CURLOPT_CONNECTTIMEOUT, 30);
 		curl_setopt($curl, CURLOPT_TIMEOUT, 30);
 
-		$content = curl_exec($curl);
-		
+		$response = curl_exec($curl);
+
 		curl_close($curl);
 
-		$line = explode("\n", trim($content));
+		if ($response) {
+			$dom = new \DOMDocument('1.0', 'UTF-8');
+			$dom->loadXml($response);
 
-		for ($i = 0; $i < count($line); $i = $i + 2) {
-			$currency = utf8_substr($line[$i], 4, 3);
-			$value = utf8_substr($line[$i], 11, 6);
-			
-			if ((float)$value < 1 && isset($line[$i + 1])) {
-				if ((float)utf8_substr($line[$i + 1], 11, 6) > 0) {
-					$value = (1 / (float)utf8_substr($line[$i + 1], 11, 6));
-				} else {
-					$value = 0;
+			$cube = $dom->getElementsByTagName('Cube')->item(0);
+
+			$currencies = [];
+
+			$currencies['EUR'] = 1.0000;
+				
+			$default = $this->config->get('config_currency');
+
+			foreach ($cube->getElementsByTagName('Cube') as $currency) {
+				if ($currency->getAttribute('currency')) {
+					$currencies[$currency->getAttribute('currency')] = $currency->getAttribute('rate');
 				}
-			}	
-						
-			if ((float)$value) {
-				$this->db->query("UPDATE `" . DB_PREFIX . "currency` SET `value` = '" . (float)$value . "', `date_modified` = '" . $this->db->escape(date('Y-m-d H:i:s')) . "' WHERE `code` = '" . $this->db->escape($currency) . "'");
 			}
+
+			if ($currencies) {
+				$this->load->model('localisation/currency');
+
+				$results = $this->model_localisation_currency->getCurrencies();
+
+				foreach ($results as $result) {
+					if (isset($currencies[$result['code']])) {
+						$from = $currencies['EUR'];
+
+						$to = $currencies[$result['code']];
+
+						$this->model_localisation_currency->editValueByCode($result['code'], 1 / ($currencies[$default] * ($from / $to)));
+					}
+				}
+			}
+
+			$this->model_localisation_currency->editValueByCode($default, '1.00000');
+
+			$this->cache->delete('currency');
 		}
-
-		$this->db->query("UPDATE `" . DB_PREFIX . "currency` SET `value` = '1.00000', `date_modified` = '" . $this->db->escape(date('Y-m-d H:i:s')) . "' WHERE `code` = '" . $this->db->escape($this->config->get('config_currency')) . "'");
-
-		$this->cache->delete('currency');
 	}
 
 	public function getTotalCurrencies() {
